requirements.txt
streamlit
numpy
opencv-python
Pillow
matplotlib
seaborn
pandas
plotly
scikit-learn
scipy